패키지명, 프로젝트명 변경 - authservice -> timerservice
