+++ server/auth-service/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'auth-service'
+rootProject.name = 'timer-service'
