--- server/auth-service/src/main/java/com/example/authservice/AuthServiceApplication.java
@@ -1,13 +1,0 @@
-package com.example.authservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthServiceApplication.class, args);
-	}
-
-}
