--- server/auth-service/src/main/java/com/example/authservice/api/controller/auth/AuthController.java
@@ -1,11 +1,0 @@
-package com.example.authservice.api.controller.auth;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@RequestMapping("/api/v1/auth")
-@RequiredArgsConstructor
-public class AuthController {
-}
