--- server/auth-service/src/main/java/com/example/authservice/api/controller/user/UserController.java
@@ -1,4 +1,0 @@
-package com.example.authservice.api.controller.user;
-
-public class UserController {
-}
