--- server/auth-service/src/main/java/com/example/authservice/api/entity/auth/AuthReqModel.java
@@ -1,15 +1,0 @@
-package com.example.authservice.api.entity.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthReqModel {
-    String email;
-    String password;
-}
