--- server/auth-service/src/main/java/com/example/authservice/api/service/UserService.java
@@ -1,4 +1,0 @@
-package com.example.authservice.api.service;
-
-public class UserService {
-}
