--- server/auth-service/src/main/java/com/example/authservice/common/ApiResponse.java
@@ -1,4 +1,0 @@
-package com.example.authservice.common;
-
-public class ApiResponse {
-}
