--- server/auth-service/src/main/java/com/example/authservice/config/security/SecurityConfig.java
@@ -1,11 +1,0 @@
-package com.example.authservice.config.security;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@RequiredArgsConstructor
-@EnableWebSecurity // Spring Security 설정을 활성화
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-}
