--- server/auth-service/src/main/java/com/example/authservice/oauth/entity/ProviderType.java
@@ -1,7 +1,0 @@
-package com.example.authservice.oauth.entity;
-
-public enum ProviderType {
-    GOOGLE,
-    KAKAO,
-    NAVER
-}
