--- server/auth-service/src/main/java/com/example/authservice/oauth/exception/OAuthProviderMissMatchException.java
@@ -1,5 +1,0 @@
-package com.example.authservice.oauth.exception;
-
-public class OAuthProviderMissMatchException extends RuntimeException{
-    public OAuthProviderMissMatchException(String message) {super(message);}
-}
