--- server/auth-service/src/main/java/com/example/authservice/oauth/info/OAuth2UserInfo.java
@@ -1,22 +1,0 @@
-package com.example.authservice.oauth.info;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes){
-        this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes(){
-        return attributes;
-    }
-
-    public abstract String getId();
-    public abstract String getEmail();
-    public abstract String getUserName();
-
-    public abstract String getImageUrl();
-
-}
