--- server/auth-service/src/main/java/com/example/authservice/oauth/info/impl/GoogleOAuth2UserInfo.java
@@ -1,31 +1,0 @@
-package com.example.authservice.oauth.info.impl;
-
-import com.example.authservice.oauth.info.OAuth2UserInfo;
-
-import java.util.Map;
-
-public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
-    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getUserName() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getImageUrl() {
-        return (String) attributes.get("picture");
-    }
-}
