--- server/auth-service/src/main/java/com/example/authservice/oauth/info/impl/KakaoOAuth2UserInfo.java
@@ -1,7 +1,0 @@
-package com.example.authservice.oauth.info.impl;
-
-import com.example.authservice.oauth.info.OAuth2UserInfo;
-
-public class KakaoOAuth2UserInfo extends OAuth2UserInfo {
-
-}
