+++ server/auth-service/src/main/java/com/gdsc/timerservice/TimerServiceApplication.java
@@ -1,0 +1,13 @@
+package com.gdsc.timerservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TimerServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TimerServiceApplication.class, args);
+	}
+
+}
