+++ server/auth-service/src/main/java/com/gdsc/timerservice/api/controller/user/UserController.java
@@ -1,0 +1,4 @@
+package com.gdsc.timerservice.api.controller.user;
+
+public class UserController {
+}
