+++ server/auth-service/src/main/java/com/gdsc/timerservice/api/entity/auth/AuthReqModel.java
@@ -1,0 +1,15 @@
+package com.gdsc.timerservice.api.entity.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthReqModel {
+    String email;
+    String password;
+}
