+++ server/auth-service/src/main/java/com/gdsc/timerservice/api/service/UserService.java
@@ -1,0 +1,4 @@
+package com.gdsc.timerservice.api.service;
+
+public class UserService {
+}
