+++ server/auth-service/src/main/java/com/gdsc/timerservice/common/ApiResponse.java
@@ -1,0 +1,4 @@
+package com.gdsc.timerservice.common;
+
+public class ApiResponse {
+}
