+++ server/auth-service/src/main/java/com/gdsc/timerservice/config/security/SecurityConfig.java
@@ -1,0 +1,11 @@
+package com.gdsc.timerservice.config.security;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@RequiredArgsConstructor
+@EnableWebSecurity // Spring Security 설정을 활성화
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+}
