+++ server/auth-service/src/main/java/com/gdsc/timerservice/oauth/entity/ProviderType.java
@@ -1,0 +1,7 @@
+package com.gdsc.timerservice.oauth.entity;
+
+public enum ProviderType {
+    GOOGLE,
+    KAKAO,
+    NAVER
+}
