+++ server/auth-service/src/main/java/com/gdsc/timerservice/oauth/exception/OAuthProviderMissMatchException.java
@@ -1,0 +1,5 @@
+package com.gdsc.timerservice.oauth.exception;
+
+public class OAuthProviderMissMatchException extends RuntimeException{
+    public OAuthProviderMissMatchException(String message) {super(message);}
+}
