+++ server/auth-service/src/main/java/com/gdsc/timerservice/oauth/info/impl/GoogleOAuth2UserInfo.java
@@ -1,0 +1,31 @@
+package com.gdsc.timerservice.oauth.info.impl;
+
+import com.gdsc.timerservice.oauth.info.OAuth2UserInfo;
+
+import java.util.Map;
+
+public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
+    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getUserName() {
+        return (String) attributes.get("name");
+    }
+
+    @Override
+    public String getImageUrl() {
+        return (String) attributes.get("picture");
+    }
+}
