+++ server/auth-service/src/main/java/com/gdsc/timerservice/oauth/info/impl/KakaoOAuth2UserInfo.java
@@ -1,0 +1,7 @@
+package com.gdsc.timerservice.oauth.info.impl;
+
+import com.gdsc.timerservice.oauth.info.OAuth2UserInfo;
+
+//public class KakaoOAuth2UserInfo extends OAuth2UserInfo {
+//
+//}
