+++ server/auth-service/src/test/java/com/gdsc/timerservice/TimerServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gdsc.timerservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TimerServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
